feat: update security layer
