--- src/main/java/dev/justedlev/configuration/SecurityConfiguration.java
@@ -1,36 +1,0 @@
-package dev.justedlev.configuration;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfiguration {
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        return http
-                .csrf(AbstractHttpConfigurer::disable)
-                .formLogin(Customizer.withDefaults())
-                .httpBasic(Customizer.withDefaults())
-                .logout(logout -> logout
-                        .logoutUrl("/logout")
-                        .addLogoutHandler(new SecurityContextLogoutHandler())
-                )
-                .authorizeHttpRequests(exchangeSpec -> exchangeSpec
-                        .requestMatchers(
-                                "/error/*",
-                                "/error",
-                                "/logout"
-                        )
-                        .permitAll()
-                        .anyRequest().authenticated()
-                )
-                .build();
-    }
-}
