+++ src/main/java/io/justedlev/msrv/ssr/configuration/SecurityConfiguration.java
@@ -1,0 +1,35 @@
+package io.justedlev.msrv.ssr.configuration;
+
+import org.apache.commons.lang.ArrayUtils;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.lang.NonNull;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfiguration {
+    @Value("${security.whitelist:}")
+    private String[] whitelist;
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(@NonNull HttpSecurity httpSecurity) throws Exception {
+        return httpSecurity
+                .csrf(AbstractHttpConfigurer::disable)
+                .formLogin(Customizer.withDefaults())
+                .httpBasic(Customizer.withDefaults())
+                .logout(Customizer.withDefaults())
+                .authorizeHttpRequests(spec -> {
+                    if (ArrayUtils.isNotEmpty(whitelist)) {
+                        spec.requestMatchers(whitelist).permitAll();
+                    }
+                    spec.anyRequest().authenticated();
+                })
+                .build();
+    }
+}
